api/requirements.txt
asgiref==3.6.0
Django==4.1.5
django-cors-headers==3.13.0
django-filter==22.1
django-resized==1.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.4.0
psycopg2==2.9.5
PyJWT==2.6.0
pytz==2022.7
sqlparse==0.4.3